+++ admin/tool/mulib/CHANGES.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## mu-5.0.2-01
+
+Release date: 09/08/2025
+
+* New modal ajax forms helper replacing dialog forms.
+* Internal refactoring.
+* Moodle 5.0.2 support.
+
 ## mu-5.0.1-01
 
 Release date: xx/xx/2025